# studyforbuddy-backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: StudyForBuddy API Documentation
  version: 1.0.0
  description: StudyForBuddy code  API backend
components:
  securitySchemes:
    BearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /classrooms/files/{id}:
    delete:
      tags:
        - "Classrooms Files"
      summary: "Delete a file by ID"
      description: "Deletes a file record and its associated file from the server"
      operationId: "deleteFile"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          description: "ID of the file to be deleted"
      responses:
        200:
          description: "File deleted successfully"
        404:
          description: "File not found"
        500:
          description: "Failed to delete file"
...